feat: Implement legacy data import
Adds functionality to import data from older versions of the application stored in local storage. This includes a new `importLegacyData` function in the AppContext and corresponding UI elements in the Settings and a new Rescue screen. This allows users to recover their data if they were using an older version of the app.